Make ModelViewer client-only (split component)
Extract client-only Three.js and @react-three code into a new
modelViewerClient.tsx and dynamically import it from modelViewer.tsx
(ssr: false). This prevents heavy WebGL libs from being included in
server bundles.
